Fixup post_event to handle styles
Tests now all pass, though the rest of the event methods need to return
styles. And that code in post_event should all be wrapped in a
transaction.

Useful documentation for understanding the DBIx::Class for all that
includes:

    - DBIx::Class::Relationship::Base
    - DBIx::Class::Manual::Joining
    - DBIx::Class::Manual::ResultSet
    - DBIx::Class::Manual::ResultSource
    - DBIx::Class::Manual::Cookbook

and the test files from the distribution:

    - t/relationship/core.t
    - t/lib/DBICTest/Schema/Track.pm

It turns out a good place to start is DBIx::Class::Manual::DocMap.

Also did some variable renaming to make the code easier to follow, fewer
generic variable names like "$data".

Also tweaked the tests to make them more declarative, which a) makes
failures easier to track and understand and b) turns the tests into a
form of documentation. Unit tests end up being somewhat repetitive and
un-DRY, but the important thing is test *failures* are obvious and easy
to understand. Unit test code is kind of disposable.