Migrate venue linkages and start on UI
The event-to-venue maps are migrated. The Venue row shows for the Event
display and is populated, but update doesn't work yet.

The Venue accordion is there and populates but isn't set up to edit any
fields yet. There are about 90 venues ATM, so we'll see if that's
scalable.

There's a couple of FIXMEs in the commit.