Fix bug with POSTing new model objects
I noticed the POST for a new style/series/caller was failing with this
message:

    Route exception:
    DBIx::Class::Storage::DBI::_dbh_execute():
    DBI Exception:
    DBD::mysql::st execute failed:
    Incorrect integer value: '' for column `schedule`.`callers`.`caller_id` at row 1
    [for Statement "INSERT INTO callers ( caller_id, created_ts, is_deleted, modified_ts, name)
    VALUES ( ?, ?, ?, ?, ? )" with ParamValues:
    0="", 1='2022-06-05T23:13:36', 2=0, 3='2022-06-05T23:13:36', 4="Some New Caller"]
    at (eval 591) line 6 in /Users/kevin/ds-lib/lib/perl5/Dancer2/Core/App.pm l. 1503

The browser is sending this POST data

    name=a.+caller4&caller_id=

Not completely sure what's going on, but I think we shouldn't be
touching the primary key field for a new object in any case--that's for
DBIx::Class to fill in.

The unit test behavior with Plack::Test seems to be slightly different.
The tests have been passing up to now because we're not sending the
caller_id or the other primary keys. If I send an empty string it
generates *almost* the same error:

    Route exception:
    DBIx::Class::Storage::DBI::_dbh_execute():
    DBI Exception:
    DBD::SQLite::st execute failed:
    datatype mismatch
    [for Statement "INSERT INTO callers ( caller_id, created_ts, is_deleted, modified_ts, name)
    VALUES ( ?, ?, ?, ?, ? )" with ParamValues:
    1="", 2='2022-04-28T02:18:05', 3=0, 4='2022-04-28T02:18:05', 5="test caller"]
    at (eval 655) line 6 in /Users/kevin/ds-lib/lib/perl5/Dancer2/Core/App.pm l. 1503

This change fixes that error as well as the first one.