Implement series defaults for new event
This adds a bunch of default values for each series, and when an event
has its series listbox changed, then the event's values change for those
defaults.

Also changed some of the sql schemas from huge VARCHARs to TEXT fields,
because for some reason the mysql on my laptop complained about them.
Had to manually back out a couple other changes that snuck in on my
laptop. Will check that dbicdump lets all this go as-is on fried.

Also filled in more requirements in the Makefile.PL as they arose.