Get rest of series default templates working.
Notes:

- In the series template popup the series name is now just some text. That's
where we'd left off the other night.

- I removed this from the Events schema:

    UNIQUE KEY (series_id, is_template)

Because that only works if the only values for is_template are 1 or NULL. After
reading this discussion
https://dba.stackexchange.com/questions/40324/only-allow-one-checked-row-in-a-column-in-sql-server
I think if we want to constrain that, we should add a separate table of
series-to-templates. And that way I can set `$new->is_template(0);` for all the
incoming events in the migration script.

- I re-ordered the Events accordion container so that the first row is "Series"
followed by the templated rows, followed by the ones the user would fill in for
an event. That'll match the actual workflow better, and more importantly, when
you're adding a new event and get down to the "Series" listbox, changing that
zeroes out all the other rows you just entered. Fixing that would be a pain,
but it just makes more sense to *start* with the Series selection anyway.

- I disabled the "Manage Templates" button when you're adding a new Series, so
you can't futz with the template until you've saved the series. We need this
because the template popup needs to get the series_id to have something to work
with. There's a bunch of spread-out code to make sure the button is re-enabled.

- When you hit "Save" on the series template popup, we left the window open so
you could look at it, but there was no indication of success, so I added an
alert(). It looks pretty crappy and there's probably a better solution.