Split start_time into start_date, start_time
...and ditto with end_time.

This is so we can have default start_times in the Series default template.

Also camps might have start date, but maybe not a start_time.

Had to add an unpretty hack to _row_to_result() because DBIx::Class was
formatting all the DATE columns as full ISO-8601 strings, which is fine for the
timestamps, but for the start_date we only want YYYY-MM-DD.

Added a filter_input() method to the Event class (empty in the base class) to add
the trailing :00 otherwise it's assumed you're talking about seconds. I was
just looking that the documentation describing that but now I can't find it.