Adjust width of UI for textareas
That 500px width we had felt too constrained to edit those textareas so I
raised it to 750px. That becomes a little problematic on a phone display, but
we don't really have a very responsive design going yet anyway.

Setting the textareas rows to `display: flex` means that lines after the first
line they won't wrap around the underneath the label there.

Setting `max-width: 40em` for the row-contents means that a long desc that's
one super-long line doesn't wrap around over the label there.

Tweaking the size of the input fields to match their contents. Again, that's
probably not very responsive.

Getting rid of that annoying "Perl Dancer" background image and adding a pretty
conic gradient.