Remove extra space in the selectbox rows
There was a tiny extra space right before the selectbox that wasn't there
before the regular input boxes that was misaligning them ever so slightly.

The problem was the "flex" on the parents, but the multi-select-wrapper is an
extra div between the spans and the parent with the "flex" on it.