Fix more betterly the textarea wrapping problem
eb96ad3b400 and eba837e1cd20 were attempts to fix the problem of the
event's rows with bit <textarea> blocks of text wrapping badly. The
first fixed that problem, but left a one-space offset, and left the
additional selectboxes in a row on the same line instead of stacked
vertically. The second one fixed the one-space problem

This fixes all the problems--just apply "flex" to the display rows with
a textarea.