Fix adding the talent for a band to an event
The `$( '#talent-for-event select' ).last();` selector was adding them
to the stuff in the series template event's popup. Making the selector
specific to the event accordion fixes that.

There's still the problem that if they just add a band, the top Talent
selectbox will be sitting there empty and can't be removed without
horking things up.

I also fixed the problem were only the first '+' next to any given list
of selectboxes did anything--the others didn't have handlers attached to
them.