Fix utf8 encoding for the app
For the UI to accept utf8 characters I had to make fixes on a number of levels:

The default charset for mysql is Latin1. So I added these to all the CREATE
TABLE statements:

    ENGINE=InnoDB DEFAULT CHARACTER SET=utf8;

Those CREATE TABLEs all then failed with this message

    Specified key was too long; max key length is 767 bytes

I fixed that per https://stackoverflow.com/a/1814594 by changing the column
size from 256 to 255 for all the VARCHARs, just assuming that would catch the
ones used as keys (and why not standardize them all?)

    -  size: 256
    +  size: 255

With MySQL now returning utf8, I needed to adjust DBIx::Class to handle it
"MySQL supports unicode, and will correctly flag utf8 data from the database if
the mysql_enable_utf8 is set in the connect options."
https://metacpan.org/pod/DBIx::Class::Manual::Cookbook#MySQL

So added this to the DBI connect arguments

    mysql_enable_utf8 => 1,

Then it turned out that Dancer2 was double-encoding the output of our
Results->format's encode_json() call. encode_json returns utf8 octets.
Apparently the dancer2 handlers expect logical perl characters which *they* can
then utf8-encode.  Lacking this call to decode-utf8 the Dancer2 handlers
double-encode the output. Maybe there's a way to tell Dancer2 to not do that?
But I added this to handle it for now:

    return decode_utf8 $json_str;

Then I added a test, which would fail without at least some of these fixes
(once I verified the UI, I didn't check every permutation in the test).

added a test, which would fails