Implement LoginSession and Get tests passing for the new auth stuff
This implements LoginMethod=session to to along with LoginMethod=google. It
checks if the LoginSession cookie is for a user in the Programmers table, but
no permissions are checked beyond that (TBD).

There's no other infrastructure around logging in with email+pw yet.

Tests now pass except for 207, whose fixing commit is on the main branch (need
to rebase)