Implement dancefinder and livecalender.
dancefinder is the HTML page at bacds.org/dancefinder and used to be
driven by scripts/dancefinder.pl. The drop-downs were populated with
separate virtual includes. This seemed silly so I reimplemented the
whole thing, both the form and the post endpoint, as
bacds.org/dance-scheduler/dancefinder

    GET /dancefinder
    GET /dancefinder-results

The implementation is mostly in bacds::Scheduler::Model::Dancefinder.
"Model" made sense to me since this is intermediating between the
controller and the storage, just like the other model classes are doing.

The SQL has been tuned to it loads everything in a single call. You can
set get_db(debug => 1) to verify that be seeing the SQL in the logs.

The templates are in:

 - views/dancefinder/index.html.tt (the input form)
 - views/dancefinder/results.html.tt (the results content)
 - views/layouts/dancefinder.tt (the wrapper)

livecalendar is the HTML page at bacds.org/livecalendar.html. It's
driven by https://fullcalendar.io and makes JSON requests to the
backend at bacds.org/livecalendar-results (used to be
scripts/dancefinder.pl, and the calendar logic was mashed together with
the dancefinder logic).

The implementation is in bacds::Scheduler, since it's just marshalling
display data after a simple call to
bacds::Scheduler::Model::DanceFinder->search_events (implemented above)

I implemented a "virtual_include" handler in bacds::Scheduler to pull in
files from e.g.  /shared/ in the production filesystem the way they do
now.