Back out NOT NULL for those three fields.
In 41fa20de5c I thought we should add NOT NULL for is_canceled,
is_template, and is_superuser. Since they're boolean fields they should
really be either 1 or 0, never null. But since the WebUI is using
checkboxes, and checkboxes don't sent a thing to the server if they're
not checked, that was causing all kinds of problems. I think we'd have
to switch them to either y/n radio buttons or to a Yes/No dropdown box
to make that happen.