Implement single-event pages
The links for the single-event pages need to be the same url base as the
series page, but with the event_id parameter:

    https://bacds.org/series/contra/palo_alto/?event_id=120

So when we're generating the single-event page links, we need to get
that SSI page's url. And when we're looking up the single-event from the
request, we need to get the query parameter from the SSI's request.