Add Auditor, Schema/Result/AuditLog
This adds a rudimentary logging to the application, as a defensive
measure so that we can attempt a reasonable answer when asked "what
happend to this event?"

Example data (edited):

    prgr  target      target_id   action      message
    ----- ----------  ----------  ---------- -------------------------------------
    1     Event       2           create      created Event "saturday night 無為"
    1     Event       1           update      name to "good times 派对"
    1     Event       1           update      styles to "CONTRA, ENGLISH"

The general "before" hook sets an "auditor" in var as a
bacds::Scheduler::Auditor object to collect the information, and the
put/post routes call $auditor->save when appropriate. (Those put/post
routes are pretty repetitive, could be rolled up.)

Also establish the "signed_in_as" var as the full $programmer object,
which simplifies things later on.