Improve UX for duplicate entry errors
Refactored Scheduler.pm and Model.pm so that e.g. get_row returns like
this:

    my ($event, $err) = $event_model->get_row($event_id);

where $err is a bacds::Scheduler::Util::Result all ready to call
->format on, with its own descriptive error string already in it.

Also rejiggered the error codes yet again. For them to make any sense
now they'd really have to be in a central registry, and that seemed like
busywork since we're not even using them, so I just changed them all
to the HTTP standards 404-Not Found and 409-Conflict.

There's still a potential problem in that "deleted" entries are still in
the unique index, so you could delete "Alice Caller", and then if you
went on to try to add her again you'd get a duplicate entry error but
you can't see the record causing the collision. If there's ever enough
actually deleted entries that it becomes a problem, we'll need to add a
UI to manage the deleted entries.

Also added some unique indexes for venues and parent orgs where I
thought they'd be helpful.

Also did a "make manifest" to update the MANIFEST which was long overdue
(not that it matters yet).