Add start_time to Dancerfinder order_by
So that events on the same day will be sorted by start_time. This came
up with the playford tea dance and contra on the same day.

The events.start_time field is defined as

    start_time TIME NOT NULL

so it's going to be some form of hh::mm:ss. But the UI doesn't
constrain it to be am/pm time or military time, and allows an empty
entry, in which case we save 00:00:00. We'll leave all that for now and
let the use cases work themselves out.  Might add a timepicker later.