Only filter talent from the event fold
Response to a bug report: when you add/edit an event, the band members
that are already in the event are supposed to be not added twice to the
event. If you look at a band and *then* go to an event, the talent
dropdowns from the band fold were interfering with the "found" check for
talent already in the event.

Solution was to constrain that search to only the event fold so that it
doesn't include the dropdowns from the talent fold.