Add team table to database and UI.
Teams are being added as their own table so that display teams can be
shown on the front page in the same way as series, and to promote
display teams generally.

In the event editing UI, you now have the option of adding teams to
events, and doing so will prompt you to add their associated styles to
the event.

Venues are no longer required, so that display team performances that
are one-off vis-a-vis location don't have to clog up the venues table.

Unit tests need to be updated.

DBIX boilerplate in subsequent commit

Still needs to be added to the end user pages for display to the
public.