Fix duplicate loading of .js files
We noticed in the browser that the .js files were each being loaded
twice:

    foo.js
    foo.js?checksum=asdfasdf

because the import() statements needed to have the checksum in them as
well. So converting the imports to dynamic imports, removing the
now-unneeded script src= tags from the html (events.js is put there by
accordions-webui.yaml and isn't loaded by scheduler.js)

Also adding a guard around that eventDisplayForm.saveHelper because the
console showed a critical error trying to push to an undefined
variable--events.js was running before scheduler.js had set up the
array. That probably means the two modules should be coordinating the
variable better, but w/ever.