Format dancefinder dates like normal people
Changing from

    Friday, 17 September, 2023

to

    Friday, September 17

and removing the year because it seems redundant. When we're in December
near the year boundary it'll be obvious which year "January" dances are
in.

And abbreviating the month names in when there's a date range, since the
repetition seemed unnecessary.

And adding an &nbsp; because the line was splitting there for BtB. Maybe
it shouldn't have spaces there at all?