Add fix-migrate-comments.pl
When I'd gotten the migration as far as 2014, I realized I'd not grabbed
any of the comments column values. I'd been building short_desc from the
leader+band, but in the old cgi-bin/calendar2.cgi, comments were
displayed along with the leader+band. So I need to go back and stick
them in short_desc for 2020-2014