Implement photo_url previews
I was starting to fill in the photo_urls for some of the talent, for the
homepage rewrite kennith and I are working on, and thought it would be
handy to have a preview of the url here in the dance-scheduler.

This adds a "preview_image: 1" bit to the accordions-webui.yml
definitions for rows where we want to preview the image.

For those rows that claim to have a preview_image we're going to add a
little div for the thumbnail. When the js loads the data for this
entity, it can grab the photo_url from the entity, stick that into the
<img> tag, and unhide the little div. And that gets hooked up to this
popup lightbox code from https://codehimblog.github.io/jquery-popup-lightbox/
which I've hacked up to do what I want.

The name of the field "photo_url" is convention. The code also looks for
"name" or "full_name" to use as the caption for the expanded image in
the lightbox. Those could be metadata-ized if we cared.

I also fixed a lowercase "<% if" to uppercase, which has been there
since 7bac0b6be.