Use same wrapper for series-page, dancefinder
This commit moves the page setup boilerplate for the series-page.tt into
a "layout", where it can be re-used by the dancefinder, and changes the
dancefinder to use it. That means the dancefinder is now
mobile-friendly. And also so I can re-use it for the archive calendar
pages. The old layout I'd used for the dancefinder was all old HTML4.0
stuff.

Also moves the "mission statement" button for desktop to the bottom of
the sidebar instead of floating around in the middle of the page like it
was.  It's easier to handle in the wrap that way and I think it looks a
little better.

It also changes the display for the navbar to flexbox so that I can push
the mission statement all the way to the bottom, and fixes the color so
there's not a white undershirt hanging out underneath the collapsible
navbar part.