Remove hard-coded "2023" from archive calendar index
Oops, I'd set up a variable with the data but hadn't referenced it in
the template.

Also changing it to eight months into *next* year because the page
does say "historical AND FUTURE" dances.

Eight months instead of twelve because it seems weird to start showing
2025 if we're only in January 2024. Eight months is sure to catch
anything way in advance.