Preserve ordering in multi-entities
Craig pointed out that when he adds talent to an event he likes to start
the list with fiddle, then other melody instruments, then accompanyment.
But the app reorders them by order of primary key in the database.

We have an "ordering" field in all the many-to-many mapping tables that
we've never used:

    # select * from event_talent_map limit 3
    +----------+-----------+----------+---------------------+---------------------+
    | event_id | talent_id | ordering | created_ts          | modified_ts         |
    +----------+-----------+----------+---------------------+---------------------+
    |      152 |         7 |        1 | 2023-01-14 22:29:49 | 2023-01-14 22:29:49 |
    |      152 |        55 |        2 | 2023-01-14 22:29:49 | 2023-01-14 22:29:49 |
    |      152 |        60 |        3 | 2023-01-14 22:29:49 | 2023-01-14 22:29:49 |
    +----------+-----------+----------+---------------------+---------------------+

This fixes it so we respect the ordering on POST and PUT and store it in
the database's "ordering" column and reflect it on GET and reflect it
in the dancefinder listings.tt.

There doesn't actually seem to be a unit test for events_with_talent,
which I think we skipped since by that point they were all boilerplate.