Add front-page styling for non-bacds events
To de-empasize them from our own events, make them slightly faded, a wee
bit smaller, and add a "Presented By" at the end. That's for the front
page of the site.

For the livecalendar, do something similar, just adding e.g. (NBCDS) to
the end to keep it more terse.

For the archive calendars, do something similar.