Weekly schedule cron mailer script
This will send a weekly mail to bacds-announce with the upcoming dances.
It's cronned up under the root user, and the "announce" header value is
installed into /var/lib/send-weekly-schedule/

The listings-text.tt wasn't thoroughly tested and may not be bug-free,
but I suspect it'll never be actually viewed. After I wrote most of it I
realized the HTML part was probably enough.