Add event_is_cancelled back to dancefinder/listings.tt
In c2a34176bd when we re-organized a bunch of listings.tt into macros,
we moved the "&mdash; CANCELED" into a macro but never called the macro.
This puts it back in so that the send-weekly-schedule.pl script can send
the unstyled HTML and the canceled events will look canceled.

The "special_class" macro is still not called. Is it even needed any
more?