Handle over-long events.name data in the app
An event had been set up with this name:

    Special traditional square dance with master caller Larry Edelman
    and master musicians Greg & Jere Canote, brought to town by Berkeley Old
    Time Music Convention.  Not to be missed!

which pushed the whole events box off the right hand side of the screen
along with the "Create New" button.

This caps the select list at the length it seems to be now and adds some
nice ellipses to show the truncation.