Correct the url for base/jquery-ui.css
Alan noticed the dance-scheduler's datepicker in the "Start Date" fields
was an unstyled pile of characters sitting on a transparent background,
basically unusuable.

It looks like in 6dd792af (just a day from exactly one year ago) we were
adjusting the jquery urls, the local fallback copies, to show the
version number, and mistakenly applied the same change to this url on
the jquery CDN. Which means this has been broken for twelve whole
months? The world wonders...