Add bootstrap to dancefinder page
Since we removed the table styling that was being used previously to
make it play nicely with the new home page.

There are a few things that are in bootstraps _reset.scss that had an
outsized impact on the rest of the old site, so they are reverted in
scheduler-pages.css; look for lines that set a style to "revert".