Handle events with no url
Alan W's Birthday Ball was set up without a series and without a custom
url, and so the links behind the address and More Info were just empty
hrefs.

He's since set up a page under events/, but in the future let's allow
listings that just exists as listings and don't have a More Info.

Fixes #27