Hide dropdown menus when navigating in-page
This was causing the default scrolling behavior for the anchor links to
miss their targets by a large margin - on mobile this was often enough
to not even show the relevent item.

Having them remain was also making the menus themselves effectively hide
the page content, which is counterproductive once you have clicked an
item.

fixes #24