Manually wrap the style buttons on the dancefinder
Having them all on one line was causing the page to be unhinged from the
left/right margins on mobile. The easy solution of adding the flex-wrap
bootstrap class is nicest except that it only rounds the corners of the
leftmost one on the first line and the rightmost one on the last line.
There doesn't seem to be a solution for that.

So just group them in fours.