Fix relative links
Since we expanded the wrapper to every page of the site the relative
links in it like "#tonight-section" have been broken. Those need to be
absolute links if we're anywhere else but on the homepage.

We really need to restructure the urls. "/" in dev should just be the
home page, not the admin js app. The difference between dev and prod "/"
vs. "/unearth" for the homepage is causing headaches.