Fix trailing slash handling in calendar urls
We've been having drama around urls like this

    https://bacds.org/calendars/2014/06

and tried fixing it on the filesystem by adding a bunch of symlinks, and
in the .htaccess file by adding a rewrite rule in an attempt to
canonicalize the with-slash version

    RewriteRule "(.*[^/])$" /calendars/$1/ [L,R=301]

However it turns out that we actually need to support four different
urls (at the moment anyway, hopefully we'll clear that up)

    https://bacds.org/calendars/2014/06
    https://bacds.org/calendars/2014/06/
    https://bacds.org/dance-scheduler/calendars/2014/06
    https://bacds.org/dance-scheduler/calendars/2014/06/

The first two represent paths on the filesystem so need the trailing
slash, and mod_dir handlese that automatically. [1]
The last two go straight to the dance-scheduler app and never see the
/calendars/.htaccess file. The "/2014/06" one was still 404'ing.

So to that end I'm just going to have the app support both slash and
slashless paths for now. Hopefully we can sort that out better at some
point. I'm pretty sure that RewriteRule above was redundant.

Some discussion of the problem here:
https://stackoverflow.com/questions/14934111/perl-dancer-trailing-slash

This adds on to https://github.com/kgoess/dance-scheduler/issues/12

[1] A "trailing slash" redirect is issued when the server receives a
request for a URL http://servername/foo/dirname where dirname is a
directory. Directories require a trailing slash, so mod_dir issues a
redirect to http://servername/foo/dirname/.