Update footer to stick to the bottom of the screen
It was showing in the middle of the screen on the dancefinder page,
because that page didn't have enough content to push it to the bottom.

This is resolved with position: fixed and bottom: 0. We put that in a
@media query so that it doesn't eat up screen real estate on mobile.
(bootstrap doesn't have responsive positional classes QQ)

This fix also exposed an error in views/dancefinder/index.html.tt (an
extra /div)