Add table for Page
See README.md for procedure to add a table.

We created the schema in pages.sql, added it to the test database via

    mysql --default-character-set=utf8mb4 \
        -uscheduler_test --password=`cat ~/.mysql-password` schedule_test \
        < schemas/pages.sql

and then ran the "dbicdump" command from README.md to generate the
Schema/Result/Page.pm file.

When we release this to production we'll need to create the table there
as well.