Change mysqldump from -uroot
Since we've got a password for mysql's root here here on fried, just
being the system root user doesn't get you in to mysql any more.
So instead we'll use a my.cnf in the directory with the
dance-scheduler private stuff, and do the mysqldump as the scheduler
user, which is better than root anyway..

Ask Kevin G. for the mariadb root password if you need it.