Correct date formatting error
Dancefinder->search_events was not sending start_date in the format
DBIx expected, which was throwing server warnings. The previous attempt
(#da28da1b0) didn't account for how that would change the results in the
case where no start_date was being sent, which was causing no dances to
show for the current day.

Now the date is having its time component truncated off before
formatting, which corrects that oversight.

Also added a new test to look for this behavior.