Sanitize series_urls on the way in
We just changed the series_url in the database, when it's a local url,
to be just /series/<style>/<thing>. Remote urls can still have a scheme
and a hostname and be whatever they want. But for our local series urls
we're going to enforce no trailing slash and all lowercase.