Add a link rel="canonical" to the series pages
As we were debating the trailing slash on the series pages we realized
we should enforce it all the way with a link rel="canonical" on the
page. This meant we needed a way to produce the canonical url for the
series pages when rendering them, and that led to a realization that the
get_full_url implementation from last time was suboptimal because it
didn't handle the test environment right.

And we're sanitizing the series_url at data entry level, in the
Model::Series subclass.

So fixed all of that.