Temporarily make the trailing slash canonical
The prior commit was leaving *some* series pages ending up in apache,
not the dance-scheduler app, and 404ing. The no-slash url was hitting a
301 MOVED PERMANENTLY redirect to the slash url because of the apache
policy:

    by default, Apache enforces a trailing slash on requests for directories
    by automatically adding one and sending a 301 redirect. This behavior
    can be disabled by setting DirectorySlash Off in the server
    configuration

And the "PERMANENTLY" part of that meant that browsers had cached the
trailing-slash version of the pages and were not even checking the
no-slash version.

So as a temporary measure, accepting that the canonical version of the
series pages ends in a slash. I've added a "?" after the final "/" to
all the entries in the .htaccess file like this:

    RewriteRule ^series/english/berkeley_wed/? /dance-scheduler/series-page?series_xid=BERK-ENGLISH [PT,L]

The series.series_url values still have no trailing slash.

We'll work this out properly the next time we meet.